Move mesh related methods from definition
